web_app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
ultralytics
pdf2image
Pillow
